#. 각 프로젝트 Packet 클래스 bool 타입 Push 로직 수정

+++ ClientSocket/Packet.cs
@@ -224,7 +224,7 @@ namespace ClientSocket
 		/// <param name="value">bool 타입 데이터</param>
 		public void Push(bool value)
 		{
-			Push(value ? 1 : 0);
+			Push((byte)(value ? 1 : 0));
 		}
 
 		/// <summary>
